flask/requirements.txt
Flask>=2.0.0
flasgger>=0.9.5
pandas>=1.0.0
numpy>=1.21.0
scipy>=1.7.0
scikit-learn==1.4.1.post1
gunicorn>=20.0.0
itsdangerous>=2.0.0
Jinja2>=3.1.0
MarkupSafe>=2.0.0
Werkzeug>=2.0.0
matplotlib>=3.0.0
mysql-connector-python==8.0.26